organize update tests
